+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Also store and retrieve copied widget config from localStorage
 * Move `Container` widget init logic from the custom-element to the main widget instance
 * Disable most `backdrop-filter` properties, they caused too many FPS issues
+* Add the ability to hide widgets from the public
 
 ## 0.2.0 (2022-11-02)